% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{create_map}
\alias{create_map}
\title{FUNCTION_TITLE}
\usage{
create_map(path = "DESCRIPTION")
}
\arguments{
\item{path}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
