% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shield_specification_helpers.R
\name{do.get.age.contact.proportions.for.model}
\alias{do.get.age.contact.proportions.for.model}
\title{do.get.age.contact.proportions.for.model}
\usage{
do.get.age.contact.proportions.for.model(
  specification.metadata,
  location,
  age.mixing.sd.mult,
  age.model,
  age.counts,
  availability
)
}
\arguments{
\item{specification.metadata}{specification.metadata}

\item{location}{location}

\item{age.mixing.sd.mult}{multiplier of the standard deviation of the age mixing model (diff_ages_partners ~ Normal(mu,sd))
used for calibration by calling \code{do.get.age.contact.proportions.for.model}}

\item{age.model}{specific age.model used to inform new partnership probabilities}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
returns a list of age contact proportions for designated group
}
