% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_remote.R
\name{find_remote}
\alias{find_remote}
\title{FUNCTION_TITLE}
\usage{
find_remote(dep, pat = Sys.getenv("GITHUB_PAT"))
}
\arguments{
\item{dep}{PARAM_DESCRIPTION}

\item{pat}{PARAM_DESCRIPTION, Default: Sys.getenv("GITHUB_PAT")}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
