% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shield_specification_helpers.R
\name{functions.sexual.contact.model}
\alias{functions.sexual.contact.model}
\title{1.functions.sexual.contact.model}
\usage{
functions.sexual.contact.model()
}
\description{
sexual contacts are charactrized via 4 components: 1)transmission probability, 2)age mixing, 3)sex mixing, and 4)race mixing
}
\references{
\file{inst/docs/sexual_contacts.docx}
}
