% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shield_specification_helpers.R
\name{get.base.initial.female.population}
\alias{get.base.initial.female.population}
\title{get.base.initial.female.population}
\usage{
get.base.initial.female.population(
  location,
  specification.metadata,
  years = DEFAULT.POPULATION.YEARS
)
}
\arguments{
\item{location}{The location for which the population data is being retrieved.}

\item{specification.metadata}{Metadata for specification.}

\item{years}{Vector of years for which to retrieve population data. Default is \code{DEFAULT.POPULATION.YEARS}.}
}
\value{
A 2D matrix showing the number of persons broken down by race (columns) within each age group (rows).
}
\description{
Generates the size of the 'female' population for the given years by calling
\code{get.base.initial.population.for.sex} for sex-specific population data.
}
\examples{
get.base.initial.female.population("C.12580", specification.metadata)
where: specification.metadata=get.specification.metadata("shield", "C.12580")
}
