% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shield_specification_helpers.R
\name{get.base.initial.population.for.sex}
\alias{get.base.initial.population.for.sex}
\title{get.base.initial.population.for.sex}
\usage{
get.base.initial.population.for.sex(
  location,
  specification.metadata,
  sex,
  years = DEFAULT.POPULATION.YEARS
)
}
\arguments{
\item{location}{The location for which the population data is being retrieved.}

\item{specification.metadata}{Metadata for specification.}

\item{sex}{The sex of the designated population ('male' or 'female').}

\item{years}{Vector of years for which to retrieve population data. Default is \code{DEFAULT.POPULATION.YEARS}.}
}
\value{
A vector of population for the given years.
}
\description{
Generates the size of the population for the given years based on sex.
}
