% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shield_specification_helpers.R
\name{get.best.guess.msm.proportions.by.race}
\alias{get.best.guess.msm.proportions.by.race}
\title{get.best.guess.msm.proportions.by.race}
\usage{
get.best.guess.msm.proportions.by.race(
  location,
  specification.metadata,
  min.age = 0,
  years = DEFAULT.POPULATION.YEARS,
  msm.proportions.by.race = c(black = 1 - 0.806, hispanic = 1 - 0.854, white = 1 - 0.848,
    other = 1 - 0.802),
  return.proportions = T,
  keep.ages = F
)
}
\arguments{
\item{location}{location}

\item{specification.metadata}{specification.metadata}

\item{min.age}{min.age #Todd: what does this mean?}

\item{years}{years #Todd: which years are these}

\item{msm.proportions.by.race}{msm.proportions.by.race #Todd: what does this mean?}

\item{return.proportions}{return proportions or frequency}

\item{keep.ages}{keep.ages #Todd: what does this mean?}
}
\value{
a 2D matrix showing the proportion of MSM by age (rows) and race (columns) #Todd: true?
}
\description{
assumes that within each county, relative risks of being MSM are as in MSM.PROPORTIONS
and total risk of being MSM is as per read.msm.proportions
}
