% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shield_specification_helpers.R
\name{get.best.guess.msm.proportions}
\alias{get.best.guess.msm.proportions}
\title{get.best.guess.msm.proportions}
\usage{
get.best.guess.msm.proportions(
  location,
  specification.metadata,
  years = 2013,
  ages = specification.metadata$dim.names$age,
  keep.age = T,
  keep.race = T,
  return.proportions = T
)
}
\arguments{
\item{location}{location}

\item{specification.metadata}{specification.metadata}

\item{years}{years #Todd: which years are these}

\item{ages}{agegroups read from specification.metadata$dim.names$age #Todd: isn't this redundant?}

\item{keep.age}{keep age #Todd: what does this mean?}

\item{keep.race}{keep race #Todd: what does this mean?}

\item{return.proportions}{return proportions or frequency}
}
\value{
a 2D matrix showing the proportion of MSM by age (rows) and race (columns)
}
\description{
generates proportion of male who are msm by race
}
