% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shield_specification_helpers.R
\name{get.female.sexual.age.contact.proportions}
\alias{get.female.sexual.age.contact.proportions}
\title{get.female.sexual.age.contact.proportions}
\usage{
get.female.sexual.age.contact.proportions(
  age.mixing.sd.mult,
  single.year.female.age.counts,
  single.year.age.sexual.availability,
  specification.metadata
)
}
\arguments{
\item{age.mixing.sd.mult}{multiplier of the standard deviation of the age mixing model (diff_ages_partners ~ Normal(mu,sd))
used for calibration by calling \code{do.get.age.contact.proportions.for.model}}

\item{single.year.female.age.counts}{number of individuals within each nominal age year for each age group}

\item{single.year.age.sexual.availability}{proportion of individuals within each nominal age year available engaged in sexual activity}

\item{specification.metadata}{specification.metadata}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
returns a list of age contact proportions for females
}
\details{
DETAILS
}
