% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shield_specification_helpers.R
\name{get.female.single.year.age.counts}
\alias{get.female.single.year.age.counts}
\title{get.female.single.year.age.counts}
\usage{
get.female.single.year.age.counts(
  location,
  population.years = DEFAULT.POPULATION.YEARS
)
}
\arguments{
\item{location}{location}

\item{population.years}{PARAM_DESCRIPTION, Default: DEFAULT.POPULATION.YEARS #Todd?}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
return counts of female in a single year
}
