% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shield_specification_helpers.R
\name{get.geographically.aggregated.race.oes}
\alias{get.geographically.aggregated.race.oes}
\title{get.geographically.aggregated.race.oes}
\usage{
get.geographically.aggregated.race.oes(
  location,
  specification.metadata,
  within.county.race.oes,
  years = DEFAULT.POPULATION.YEARS,
  as.functional.form = T
)
}
\arguments{
\item{location}{location}

\item{specification.metadata}{specification.metadata}

\item{within.county.race.oes}{observed to expected ration for mixing by age within each county}

\item{years}{PARAM_DESCRIPTION, Default: DEFAULT.POPULATION.YEARS #Todd???}

\item{as.functional.form}{PARAM_DESCRIPTION, Default: T #Todd???}
}
\value{
todd??
}
\description{
within.county.race.oes\link{to,from} is how much more likely someone it is for a person of race to to have a partner of race from, relative to race from's population prevalence
}
