% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shield_specification_helpers.R
\name{get.heterosexual.male.single.year.age.counts}
\alias{get.heterosexual.male.single.year.age.counts}
\title{get.heterosexual.male.single.year.age.counts}
\usage{
get.heterosexual.male.single.year.age.counts(
  location,
  specification.metadata,
  population.years = DEFAULT.POPULATION.YEARS
)
}
\arguments{
\item{location}{location}

\item{specification.metadata}{specification.metadata}

\item{population.years}{PARAM_DESCRIPTION, Default: DEFAULT.POPULATION.YEARS #Todd???}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
To determine the proportion of the population that falls into specific age buckets
}
