% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shield_specification_helpers.R
\name{get.proportion.msm.of.male.by.race.functional.form}
\alias{get.proportion.msm.of.male.by.race.functional.form}
\title{get.proportion.msm.of.male.by.race.functional.form}
\usage{
get.proportion.msm.of.male.by.race.functional.form(
  location,
  specification.metadata
)
}
\arguments{
\item{location}{location}

\item{specification.metadata}{specification.metadata}
}
\value{
??? #Todd??
}
\description{
Generates proportion of male who are msm by race
}
