% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shield_specification_helpers.R
\name{get.sexual.availability}
\alias{get.sexual.availability}
\title{1-get.sexual.availability}
\usage{
get.sexual.availability()
}
\value{
1D vector with proportion of people in each age bucket that are sexually available
}
\description{
Determines the proportion of people in each age bucket that are sexually available
}
\details{
The model reflects an increase in sexual activity starting from age 13, reaching 100\% at
ages 20 to 64, and gradually tapering off until age 85, the final age group.
}
