% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{install_remotes}
\alias{install_remotes}
\title{FUNCTION_TITLE}
\usage{
install_remotes(obj, ...)
}
\arguments{
\item{obj}{PARAM_DESCRIPTION}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[remotes]{install_github}}, \code{\link[remotes]{install_cran}}
}
