% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shield_specification_helpers.R
\name{oes.to.proportions}
\alias{oes.to.proportions}
\title{oes.to.proportions}
\usage{
oes.to.proportions(oes, population)
}
\arguments{
\item{oes}{oes betwen groups}

\item{population}{popualtion count in each group}
}
\value{
todd??
}
\description{
trasforming oe values to proportions of mixing with other groups (sum to 1)
}
