% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_sha.R
\name{repo_sha}
\alias{repo_sha}
\title{FUNCTION_TITLE}
\usage{
repo_sha(repo, pat = Sys.getenv("GITHUB_PAT"))
}
\arguments{
\item{repo}{PARAM_DESCRIPTION}

\item{pat}{PARAM_DESCRIPTION, Default: Sys.getenv("GITHUB_PAT")}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
