% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shield_specification_helpers.R
\name{sexual.oes.to.contact.proportions}
\alias{sexual.oes.to.contact.proportions}
\title{sexual.oes.to.contact.proportions}
\usage{
sexual.oes.to.contact.proportions(race.sexual.oes, race.population.counts)
}
\arguments{
\item{race.sexual.oes}{oes}

\item{race.population.counts}{number of people by race}
}
\value{
todd??
}
\description{
trasforming oe values to proportions of mixing with other races (sum to 1)
calling code \code{oes.to.proportions}
}
