% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shield_operations.R
\name{shield_describe_emergency_contact_settings}
\alias{shield_describe_emergency_contact_settings}
\title{Lists the email addresses that the DRT can use to contact you during a
suspected attack}
\usage{
shield_describe_emergency_contact_settings()
}
\description{
Lists the email addresses that the DRT can use to contact you during a
suspected attack.
}
\section{Request syntax}{
\preformatted{shield$describe_emergency_contact_settings()
}
}

\keyword{internal}
